+++ rule_runner_label.go
@@ -11,6 +11,9 @@ const (
 	compatInvalid                   = 0
 	compatUbuntu2204 runnerOSCompat = 1 << iota
 	compatUbuntu2404
+	compatUbuntu2604
+	compatXcode270
+	compatXcode270XL
 	compatMacOS140
 	compatMacOS140L
 	compatMacOS140XL
@@ -26,6 +29,7 @@ const (
 	compatWindows2025
 	compatWindows2025VS2026
 	compatWindows11Arm
+	compatWindows11VS2026Arm
 )
 
 // https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners
@@ -36,15 +40,20 @@ var allGitHubHostedRunnerLabels = []string{
 	"windows-2025-vs2026",
 	"windows-2022",
 	"windows-11-arm",
+	"windows-11-vs2026-arm",
 	"ubuntu-slim",
 	"ubuntu-latest",
 	"ubuntu-latest-4-cores",
 	"ubuntu-latest-8-cores",
 	"ubuntu-latest-16-cores",
+	"ubuntu-26.04",
+	"ubuntu-26.04-arm",
 	"ubuntu-24.04",
 	"ubuntu-24.04-arm",
 	"ubuntu-22.04",
 	"ubuntu-22.04-arm",
+	"xcode-27",
+	"xcode-27-xlarge",
 	"macos-latest",
 	"macos-latest-xlarge",
 	"macos-latest-large",
@@ -82,10 +91,14 @@ var defaultRunnerOSCompats = map[string]runnerOSCompat{
 	"ubuntu-latest-4-cores":  compatUbuntu2404,
 	"ubuntu-latest-8-cores":  compatUbuntu2404,
 	"ubuntu-latest-16-cores": compatUbuntu2404,
+	"ubuntu-26.04":           compatUbuntu2604,
+	"ubuntu-26.04-arm":       compatUbuntu2604,
 	"ubuntu-24.04":           compatUbuntu2404,
 	"ubuntu-24.04-arm":       compatUbuntu2404,
 	"ubuntu-22.04":           compatUbuntu2204,
 	"ubuntu-22.04-arm":       compatUbuntu2204,
+	"xcode-27":               compatXcode270,
+	"xcode-27-xlarge":        compatXcode270XL,
 	"macos-latest-xlarge":    compatMacOS150XL,
 	"macos-latest-large":     compatMacOS150L,
 	"macos-latest":           compatMacOS150,
@@ -106,9 +119,10 @@ var defaultRunnerOSCompats = map[string]runnerOSCompat{
 	"windows-2025-vs2026":    compatWindows2025VS2026,
 	"windows-2022":           compatWindows2022,
 	"windows-11-arm":         compatWindows11Arm,
-	"linux":                  compatUbuntu2404 | compatUbuntu2204, // Note: "linux" does not always indicate Ubuntu. It might be Fedora or Arch or ...
-	"macos":                  compatMacOS260 | compatMacOS260Intel | compatMacOS260L | compatMacOS260XL | compatMacOS150 | compatMacOS150Intel | compatMacOS150L | compatMacOS150XL | compatMacOS140 | compatMacOS140L | compatMacOS140XL,
-	"windows":                compatWindows2025VS2026 | compatWindows2025 | compatWindows2022 | compatWindows11Arm,
+	"windows-11-vs2026-arm":  compatWindows11VS2026Arm,
+	"linux":                  compatUbuntu2604 | compatUbuntu2404 | compatUbuntu2204, // Note: "linux" does not always indicate Ubuntu. It might be Fedora or Arch or ...
+	"macos":                  compatXcode270 | compatXcode270XL | compatMacOS260 | compatMacOS260Intel | compatMacOS260L | compatMacOS260XL | compatMacOS150 | compatMacOS150Intel | compatMacOS150L | compatMacOS150XL | compatMacOS140 | compatMacOS140L | compatMacOS140XL,
+	"windows":                compatWindows2025VS2026 | compatWindows2025 | compatWindows2022 | compatWindows11Arm | compatWindows11VS2026Arm,
 }
 
 // RuleRunnerLabel is a rule to check runner label like "ubuntu-latest". There are two types of